# Ques_backend/docs/api/api_specification.yaml
openapi: 3.0.3
info:
  title: Ques Backend API
  description: |
    Complete API specification for the Ques application backend.
    Includes AI-powered user discovery, chat system, swipe mechanics, payments, and more.
  version: 1.0.0
  contact:
    name: Ques Backend Team
    email: backend@ques.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ques.app
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /contacts:
    get:
      tags:
        - Contacts
      summary: Get contact list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Contacts retrieved

    post:
      tags:
        - Contacts
      summary: Add contact
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contactId:
                  type: string
                notes:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Contact added

  # User Settings endpoints
...